redirect fix
