+++ spec/requests/students_spec.rb
@@ -6,12 +6,12 @@ RSpec.describe "Students", type: :request do
     it "gets a list of all students" do
       user = User.where(email: 'test@example.com', password: 'password', password_confirmation: 'password')
  
-      user.students.create(
+      Students.create(
         name: "Daniel LaRusso",
         notes: "Became a black belt in like 3 days",
         rank: "Black Belt",
         image: "https://mediaproxy.salon.com/width/1200/https://media.salon.com/2015/08/daniel-the-bully-karate-kid.jpg",
-        user_id: user.id
+        # user_id: user.id
       )
       get '/students'
